Unit3/Ragchatbot/requirements.txt
numpy==1.24.4
streamlit==1.31.1

langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.53

langchain-groq==0.1.5
langchain-cohere==0.1.5

chromadb==0.4.24
pypdf==3.17.4

python-dotenv
tiktoken
huggingface-hub==0.20.3
